Move private key reuse out of server loop
As the server list changes we can get into the situation where the first
host doesn't have a corresponding config file yet, so when we try to look
up the old privkey for that host we'll come up empty and generate a new one
instead.

Now we just look for the first privkey we come upon across all config
files.

This also removes the mullvad configs before writing the new ones to prune
removed servers. This does unfortunately make it more difficult to compate
the old pubkeys of servers so we stop warning about that for now.